notification-services/requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
async-timeout==3.0.1
asynctest==0.13.0
atomicwrites==1.3.0
attrs==19.3.0
black==19.3b0
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
firebase-admin==3.2.0
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.6.0
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
grpcio==1.24.1
grpcio-tools==1.24.1
httplib2==0.14.0
idna==2.8
importlib-metadata==0.23
Jinja2==2.10.3
MarkupSafe==1.1.1
marshmallow==3.2.1
more-itertools==7.2.0
msgpack==0.6.2
multidict==4.5.2
packaging==19.2
pluggy==0.13.0
protobuf==3.10.0
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyfcm==1.4.7
pyparsing==2.4.2
pyTelegramBotAPI==3.6.7
pytest==5.2.1
pytest-asyncio==0.10.0
python-http-client==3.2.1
pytz==2019.3
requests==2.22.0
rsa==4.0
sendgrid==6.1.0
six==1.12.0
toml==0.10.0
uritemplate==3.0.0
urllib3==1.25.7
wcwidth==0.1.7
yarl==1.3.0
zipp==0.6.0
